% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendertable.R
\name{rendertable}
\alias{rendertable}
\title{custom renderTable function}
\usage{
rendertable(expr, ..., env = parent.frame(), quoted = FALSE, func = NULL)
}
\arguments{
\item{expr}{a list with named \code{data.frame} objects in it}

\item{...}{everything else like in \code{\link[shiny]{renderTable}}}
}
\description{
can render a list of tables
}

